requirements.txt
streamlit>=1.37.0  # 需 1.37+ 以支持 st.fragment（标签延迟加载、切换视频）
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.0.0
bcrypt>=4.0.0
httpx==0.27.2
mem0ai>=1.0.0
